Adjust invisibility costs and contamination
To simplify the effect and its costs, remove the contamination over time,
and reduce the initial contamination for activating the ego. As with
Irradiate, the darkgrey to lightgrey threshold is now also the threshold at
which activating the ability again might cause yellow contamination.
Increase the drain cost slightly to further penalise using it at lower
success rates, and to make it less likely that the kills during the effect
can quickly counteract the draining.

Also remove the contamination from invisibility beams, since wands of
invisibility are long gone.