Remove the Orb's effect on contamination
Contamination is now much less a common status that players need to manage,
and more a rare effect from miscasts and specific abilities/items. As such
it no longer makes as much sense for the Orb to have a special (and fairly
minor) interaction with it. If any special interaction is still desired,
it'd be better to increase contamination received rather than to just
reduce dissipation slightly.