Simplified Divine Vigour
The effectiveness and duration of Divine Vigour become stronger in proportion to the Invocation skill level. However, the usefulness of this power fluctuates excessively depending on the skill level. It's also annoying to guess how much the effect increases depending on the skill level. Therefore, I would like to simplify this power like Heroism.
I plan to fix the effect of this force to 'MaxHP/MaxMP 30% increase' and lower the cost slightly.