Rework the autumn katana
Now, the sword sometimes uses manifold assault when attacking an enemy,
instead of having clarity.

Due to issues of setting time_taken and preventing re-entry the katana
duplicates some of the functionality of cast_manifold_assault instead of
simply re-using the function. However! There are two great upsides to
this implementation: monsters can use the katana (no special flag) and
players can use it with manifold assault to get one layer of recursion
(and a rare message).

[ Committer's note: Rewrote the artefact function, tweaked the
  implementation, revised the commit message. Closes #2234 ]