fix: correctly interpret `~` in crawl_dir etc
In most cases, this was just getting used for its string value, leading
to the creation of subdirectories named `~/`, pretty amazing. For some
uses of `SysEnv.crawl_dir`, `_resolve_dir` was getting correctly called
on subdirectories, leading to probably inconsistent directory patterns,
but I'm not sure of the details. The directory handling code here is
very confusing.