fix: resequence rc directory setting
Previously, when setting `crawl_dir` in an rc, the following sequence of
events would happen:

1. crawl_dir gets set to a default value (e.g. a ~/Library directory on
   mac.)
2. reset_options initializes save_dir based on this crawl_dir.
3. crawl_dir changes (stored in static SysEnv)
4. a. maps are loaded, des cache is built in the save_dir set in 2
   b. game menu is also initialized from this save_dir
5. player starts game, triggering another reset_options. Now the
   save_dir is set based on the crawl_dir in 3, instead of 1. This is
   what we want, but:
6. No des cache in the current save_dir, so the game crashes any time
   levelgen is triggered

The main change in this commit is to do the path resets from step 2 any
time `crawl_dir` is set, so that the save_dir can't get out of sync like
this. There's a bunch of refactoring that I did while working on this,
some of which didn't end up being necessary but I kept because it seemed
cleaner. This commit also explicitly rechecks for a data directory if
crawl_dir changes (since the data directory could be under crawl_dir,
though it's not in most cases I think). The overall directory handling
code is pretty insane but it's covering so many special cases across
multiple platforms that it's a bit hard to refactor safely.

This commit also adjusts the documentation. Partly this is to indicate
this change (since setting save_dir before crawl_dir gets overridden),
but partly this is just to make it more accurate. Unfortunately I
couldn't really even figure out how the defaults are interpreted outside
of mac, so more accurate -> more vague in this case. Someone could try
to make this more specific, especially for the defaults for `crawl_dir`
and `macro_dir` -- it would be best to explicitly list the defaults for
each of win/mac/linxu. I think based on the report in #2192 it might use
the current directory on linux for crawl_dir?

Resolves #2192.

(cherry picked from commit 381dff06d7c1ae145bb1e2932c451477f76b83b3)