fix: generalize some relayout code from rings
In the past, it seems that evokable abilities were mostly confined to
rings, and so the code that checked for ability changes was only done
when putting on a ring. This commit generalizes the (extremely brute
force) check so that it happens whenever any equip is added or removed
-- so this should handle scarfs of invis, unrands that have abilities,
etc.

Resolves #1754

(cherry picked from commit bdc701c5260bdb3288f8ac54e866a279d0fa0830)