feat: refactor debug build autowiz with an options default
Previously, when working with a debug build, to get a truly non-wizard
character you had to rebuild entirely (either switch to a different
build, or comment out the wizard mode setting in startup.cc). Suppress
wizmode is still not perfect, and it is useful to be able to start a
regular character from time to time.

This commit reimplements the auto-wiz feature so that instead of a debug
build simply forcing wizmode directly on character creation, a debug
build defaults the `wiz_mode` option to `yes` instead of `no`. (N.b.
this doesn't override the logic in a dgamelaunch build.) This lets a
debug build respect `wiz_mode=no` (or `never`) in an rc file.