fix: handle button commands in map view (local tiles)
There were two interrelated bugs here: command handling was simply
unimplemented (a key was returned by tiles code, but the widgetified
version of the map UI ignored it), and any clicking (including on
buttons) was interpreted as dungeon clicking, moving the cursor to a
clamped map position. I guess this probably just fell through the cracks
when the map view was widgetified.

At least because of the new API, we can now pass the command directly to
the map view, rather than needing to pass a key. I've implemented this
with an unappealing dynamic_cast, someone may want to do something nicer
here.

Resolves #2087