fix: disable most tiles buttons in map view
All sorts of stuff was semi-clickable in map view, e.g. you could cast
spells or use abilities via local tiles buttons. This commit tries to be
more consistent about it: anything that works in KMC_LEVELMAP can work,
and anything that doesn't, can't. It's possible that this disables some
useful stuff that can work (even if it isn't designed to) that tiles
players will be used to? We'll see if there are complaints...