fix: set ui_root.needs_paint after a resize
This resolves #2218. The proximate issue seems to have been that
resizing with ctrl-command-f both exits the menu and triggers a resize
event, which is handled immediately when pump_events is next called
during; this is immediately on displaying the game menu. Something about
being the first event was messing up the repaint. I'm not sure what.
Beyond that, I think there is still something weird going in the ui.cc
code for this case: in my testing on a large monitor I see a brief flash
of a very strangely resized window before the new size takes hold.
Unfortunately I don't understand ui.cc well enough to have any sense of
what might be doing this.