fix: don't fire tracers on LRD validity check
This is a somewhat nasty bug that triggers when returning to the main
menu after playing; because there is still a map kicking around, when
trying to check if LRD is valid, the LRD setup code fires off some
tracers in that map, leading to bad memory access. I'm 90% sure this is
without manually inserting the ASSERT here. (But llvm is usually
"better" about not crashing on bad memory accesses.) My approach here
was pretty brute force: I didn't see how to make the LRD setup code safe
so I just force-validated it. Even on the clean startup case, it's still
interacting with suspect memory regions, even if it never fires tracers.

Possibly env stuff ought to be cleared up in _reset_game() so that the
global state during startup steps is more predictable.  In general, I
wonder about the rest of the monster spell setup code triggered in this
init step...

Resolves #2168 (probably)