Add static targeters for many abilities
Making abilities behave more consistently with regards to targeting is a
pretty substantial project, but this is a start for many of the simpler
abilities. Currently this only supports static targeters - some limited
support for regular targeters has been added, but more work would be
required to fully support regular targeters (such as more ability flags to
indicate how an ability should be targeted).

As with spells, a number of abilities with either full-LOS or limited
radius effects are added to the default force_ability_targeter option.