Remove Sea of Fire
Ignis's Sea of Fire was a very visually cool effect, but it had no clear
purpose. It wasn't clear when players should use it - the most common
experience was for players to use it once, get trapped by it, feel extreme
regret, and never use it again.

Let's replace it with something with a greater clarity of purpose.