refactor: remove an unused build conditional
As far as I can tell this #define and its conditionals have never been
used by anyone. The only use case anyone has come up with is shared
accounts (geekosaur), but if this is an actual use case that someone
wants I'd rather implement it in a different way (e.g. CLO or just a
regular rc option).