New Ignis ability: Fiery Armour
Replacing Sea of Fire for the 'cheap ability' slot, Fiery Armour is
a proactive defensive ability, complementing Foxfire Swarm (a reactive
offensive ability) and Rising Flame (a reactive escape). When used,
Fiery Armour gives +7 AC and 2d4 fire damage to melee attackers. It
lasts for somewhere around 120 turns, to help it feel better compared
to Foxfire Swarm - the latter straight up wins a fight, so it's nice
if Fiery Armour can help with a few fights. (It also provides an
incentive not to dawdle, which is fun!)

There's a special message when you use it with Ozo's Armour, Icemail,
or Ice Form, as a bit of a lampshade.