New gloves ego: "infusion"
A riff on the old Charms spell of the same name, gloves of Infusion
give +0,+4 slaying (that is, +0-4 pre-AC damage per hit) on melee
attacks, in exchange for 2 MP spent per hit. This will be appealing
for many characters, but will you regret being stuck at 0 MP in a
tight spot? Maybe!

Numbers are very preliminary. In testing, my main concern was whether
it'd cause too much resting, but it didn't seem to be worse than
'caster' characters already have it. Excited to hear feedback from
others.