feat: api to track GameOptions modified by rc settings
This provides a way of checking whether a GameOption subclass was
explicitly modified by a player's rc. The eventual point of this to
ensure that explicit settings can override an autosaved preference.

Obviously, because this is tied to use of GameOption, it does not work
on custom-parsed options handled in read_option_line, nor on options
that are stored on the lua side. It also doesn't track any changes to
member variables of game_options made directly.