feat: save `default_manual_training` in `.prf` files
This is more of a proof of concept commit than a complete implementation
of this idea that can be generalized immediately to tons of options, but
it seems to work well enough and provides a path towards generalized
option serialization in .prf files.

I chose `default_manual_training` because it already has a UI element
that players might reasonably expect to be saved, it involves no custom
parsing (just a simple `BoolGameOption`), and because if something goes
wrong with the .prf files for it, it doesn't seem too disastrous
(because the it's ignored after game start). The .prf files are already
used to store newgame options, and are parsed using game_options reading
code already, so this change was a matter of figuring out how to
coherently write non-newgame options, and pull them in to the Options
static in a sensible way.

Some notes:

* If a player has an option explicitly set in their rc file, it will
  override the .prf version. This part is written in a general way so it
  will extend to new serialized options.
* The general code only works for options implemented using a GameOption
  object. Getting it to work for others will need a more effort,
  including custom handling in the `merge` function and changes to the
  tracking of what was loaded from an rc. Possibly more options should be
  converted to `GameOption`s rather than more custom handling being added.
* Serialization code is extremely not general, this would be easy +
  tedious to generalize so I will leave that for when more options are
  added.
* A bunch of boilerplate code is moved to game-options.h. This stuff
  definitely has the feel of something that ought to be templated, but
  this doesn't work with the polymorphism needed for GameOption in any
  way that I have yet found. This commit also adds a generic function to
  set one GameOption object from another, with a somewhat ugly
  implementation. (Also should be templated, but I haven't come up with
  a way yet.)
* Most other options will need UI added to do something like this.
* It's possible that we should add a game menu command that lets you
  reset the .prf file before adding too much more to this.