feat: mark identify scrolls as useless once everything is identified.
Create a "have identified all potions and scrolls" prop. Scrolls of
identify are treated as useless if it is set. It's updated on game load
if the version has changed, and also whenever set_ident_type() changes
something.

This also refactors several checks for complete identification to check this
prop.

[ Committer's note: Closes #2244. Added Mu special casing. ]