fix: don't force picking up of newly IDd equipment
This just re-uses the logic that was being used for wands, which also
avoids the edge-case where if you floor-id something without using
autopickup triggering (like when AP is disabled) then it was possible to
have the item ID'd, marked as don't-pickup in `\`, and still picked up
due to NEEDS_AUTOPICKUP_KEY set in its props.

The old code that led to this logic seems to imply there was an
intentional hack for "previously-unknown items", but in testing, I found
no unexpected behavior with the new version.