feat: nerf flame wave damage
Flame Wave did calcdice_calculator<3, 10, 2, 5> damage for each blast
(at a total cost of 6mp over 3 turns)
Compare to some other level 4 damage spells:

- Hailstorm calcdice_calculator<3, 10, 1, 2>
  (more awkward targeting, costs 12 mp to fire for 3 turns)

- Mystic Blast calcdice_calculator<2, 6, 1, 3>
  (also has knockback (for possible knockback damage), easier targeting;
   still 12 mp to fire for 3 turns)

- Sticky Flame calcdice_calculator<2, 3, 1, 12>
  (also sets them on fire for a time, doing 2d4 per turn extra damage;
  initial sticky flame damage plus 2 turns of fire damage is the
  calcdice plus 4d4 for 4mp over 3 turns)

At equal training, flame wave can do as much as half the damage of 3
casts of starburst, for 1/3rd the mp. There's targeting awkwardness and
the hazard of being adjacent but this damage is just really really good.
With 3 dice it is also consistent in its damage output.

So let's nerf it to calcdice_calculator<2, 8, 1, 5>.