fix: only merge options different from default (ardl)
The feature in 843f935 was implemented by adding merge functionality to
the game_option class. Unfortuantely, this led to defaults being loaded
from init files (instead of included in the binary) being clobbered as
they were not set in the temprorary perf-loaded options.

This changes the implementation of merge to only merge entries that were
loaded in the source game_options object.

The buggy behavior was most notable in how it interacted with the
auto_exclude default.