fix: draw complete travel trail when autoexploring (#2081)
After 14faa605, the travel trail sometimes doesn't have the last step.
It happens when autoexplore is interrupted by a monster coming into
view or by the player noticing a new item pile.

Previously, we removed the running delay from the delay queue only after
adding the last visited tile to the trail.

We stop running and remove the delay earlier now, so we need an extra
check after the code added in 14faa605.