Nerf merfolk impalers/javelineers
Javelineers are the deadliest enemy in Shoals, and it's a bit odd that
they are. A common nice dynamic in Crawl is that the closer you get to
an enemy, the more dangerous they are (so closing in to melee is a risk).
Occasionally, we toss in a curveball and have enemies that are neutered
in melee, eg centaurs. But javelineers are just consistently deadly at
all ranges. That's not awful, but it seems like in melee, other enemies
should be nastier.

So, drop javelineer damage by 5 (from 27 [with javelin] to 22), and remove
2 EV (14 -> 12) and 5 avg HP (78 -> 73).

Also pull back on impalers slightly, as a belated response to melee enemy
buffs from weapon delay changes. (Damage with trident down from 34 to 31.)

This might make Shoals a bit closer to Snake and Swamp in overall difficulty
(varying with character, of course), which seems good.