feat: black torch: zombie piety
Move Yred to an alternative piety system, where abilities are paid by
offereing up some of the player's reaped souls.

Currently this uses Zombies-recently-in-LOS for piety, the aim being a
medium-time-scale piety system that builds up over a single floor, but
resets on floor transition. Zombies on the floor could also be used, but
there are information leak issues with that, plus it encourages closet
stashing. The goal is to encourage the player to keep their harvest
close at hand!

Reaped monsters live on the floor they were created on and disappear
when the player leaves the floor (by any means, portal vaults and
shafts both count). This gives Yred a medium-scale piety system, lasting
longer than Usk but much shorter than a conventional god.