Make Swamp a bit more magical
Give out guaranteed wands to swamp-exclusive Eleionomae and Fenstrider
Witches - on average, about 5 wands per Swamp, about 10% of the total
generated over the course of a game. This is intended to be something
of a counterpart to the ammo generation in Shoals, and to complement
the earlier reduction of wand charge generation.

We can increase these numbers farther if need be.