Reduce the average number of items in a shop
With the pool of permanent items slowly decreasing over the years,
shops have shown a noticeable increase in redundancy with permanent
items, such as jewellery stores selling multiple copies of a particular
amulet type. In addition, shops can generate with up to 16 items,
which creates unnecessary UI clutter. Fix this issue by reducing the
average number of items in a shop from 4 + 3d12/3 to 4 + 3d8/3, reducing
inventory by 2 items on average.

This change also counterbalances a gradual increase in the quantity
of consumables available, possibly due to the removal of food and
gadget shops.