Don't highlight spells Vehumet offers if you've read a book about them.
This changes the meaning of "blue spell at the top of the list" from "spell Vehumet is offering you" to "spell which is only available because Vehumet is offering it to you". Vehumet can withdraw the offer, but you cannot lose the book.

This change doesn't affect the religion screen.