Qaz's power cost adjustment
In early games, the risk is greater than the benefits of worshiping Qaz. His signature characteristic, the unstoppable noise, is a fatal disqualification, but the powers given are too expensive to overcome the ordeal and require long training to be strong. So, I have somewhat eased the cost of power. I hope the players will be able to overcome the hostile environment more actively.