Refactor many ability fail checks
Abilities should always be targeted before their fail check, so that they
can consistently be cancelled before committing to activating the ability.
This commit fixes this issue for a large number of abilities, and removes
unnecessary fail checks for abilities that should never fail. A small
number of abilities still need further refactoring to move the fail check
to take place pre-targeting.