fix: min panel scale to 50%, not 10%
With 0.1, it becomes hard (on my display at least) to get back to the
panel by right clicking. 50% is about the smallest I could imagine it
being usable at, and it makes it possible to get to the settings menu to
reset this value. This may need further tweaking. Note that nothing
prevents you from setting a lower value in rc if this is useful in some
configuration I'm not imagining. (There's a different way to hide it
altogether.)