feat: use focus-trap for action panel settings popup
This fixes a whole bunch of issues with keypresses going the wrong
place, focus not being returned after this closes and just generally
being weird, etc. There are still various issues with this commit that
are indicated in the comments, where basically (to me at least) browser
focus can be extremely mysterious. But the only user-facing one is that
I couldn't get tabbing to the buttons to work.

Also, rename some html elements.