feat: add optional weapon argument to monsterinfo.target_weapon
Adds an optional `weapon` argument to the Lua
`monsterinfo.target_weapon` function. If `weapon` is omitted, it shows
the targeting description accompanying an unarmed attack. If `weapon`
is provided, it shows the targeting description for an attack with that
weapon.

This changes usage of the function slightly (formerly, the function
returned only the description for your current weapon), but the function
is new in this trunk cycle.

[ Committer's note: Tweaked to also resolve a TODO in describe.cc.
  Closes #2275 ]