feat: refactor eleionoma and fenstrider wands
2dcde036d21 gave eleionoma and fenstirder witches guaranteed wands.
However, it didn't give them the ability to use the wands. This was on
purpose, the wands were meant as loot for the player, not an extra
threat on already mid and high tier monsters.

Placing the items into monster inventory was a bit confusing, players
expect monsters carrying wands to be capable of using them; especially
with the UI efforts put in to announce monster wands.

Since these are meant as pure loot, this commit refactors them to drop
on death through the same code path that drops dragon hides. This
roughly halves (corpse rng) the rate of wand generation in 2dcde036d21,
so the charges dropped are doubled to compensate.