refactor: black torch: treat bound souls as their own species
There were a bunch of strange special cases for spectral things intended
to handle edge cases for bound souls but not show up on regular
spectrals because of other checks or luck. This changes the enslaved
soul to be its own monster type, still of species spectral thing, and
simplifies or makes explicit the relevant checks.

This commit also allows bound soul stats to be adjusted independent of
spectral things; the EV modifier could be changed in the future. For
now, to buff the bound soul, it gets full HP so as to be less fragile.