rename: change Yred abilities to not be explicit slavery
Casual use of slavery-as-entertainment in crawl has been on its way out
as poor taste for a while. Yred was one of the last hold outs, this
rework is a good opportunity to make a change. Instead of dipping into
the metaphysics of whether or not "enslaving souls" is the same as
chattel slavery, let's just be a bit more vague in the flavour of Yred.