fix: improve `.prf` file naming/location on DGL builds
Two changes, should affect DGL builds only:

1. Set the filename correctly. After game start, it turns out that
Options.game.name is the empty string, resulting in the hilarious
behavior of everyone on a DGL_STARTUP_PREFS_BY_NAME build sharing
a single prefs file named `start--ns.prf`.

2. For a DGL_STARTUP_PREFS_BY_NAME build, put prefs in a subdirectory of
the shared directory, rather than the saves directory. The reason for
this is that the saves directory on a DGL setup (and maybe any setup
using this flag) is versioned, and so the .prf file gets wiped
out/ignored on any trunk version change. (An alternative would be having
DGL copy the files, but this would be too painful to implement.)