fix: correct hidpi rendering for the action panel (PleasingFungus)
This code assumed that the `renderer` object was pre-adjusted for the
device to pixel ratio (which is how it's done elsewhere), but this is
not so straightforward to do for the action panel (though it might be
cleaner if it could be done). So, do explicit scaling for the device
ratio, and remove explicit scaling for mouse cursor position. Should
have no impact on non-hidpi/retina devices.