fix: handle db source files with mtime 0
These don't occur normally in the wild, but sometimes people doing
reproducible packaging or the like will set all mtimes to 0. When this
happens some old db code that treated 0 as an error condition would
produce a fairly incomprehensible ASSERT. I suspect this is what is
behind:

  https://crawl.develz.org/mantis/view.php?id=11956&nbn=1#bugnotes
  https://github.com/crawl/crawl/issues/2276

and the like, though this is not something I've been able to test. I've
also rewrote the ASSERT to produce something more like a real error
message, though I think it should be much harder to trigger now.

Note to packagers: this commit keeps things from crashing, but
distributing files with mtime 0 is *really not recommended* unless you
are sure to never use a shared db cache directory. If you do, this
commit will severely break having multiple versions installed (and I
don't think the errors will be any less cryptic that this one). It
should meet a typical reproducible build desiderata to use the official
file times for db (and most dat/ files) as these are not generated at
build time -- they are game data files.