Adjust some Qazlal ability costs
To help Qazlal out a bit early on, reduce the MP and piety cost of Upheaval
slightly, and also reduce the MP cost of Elemental Force. Also increase the
MP cost of Disaster Area by 1, to make it a tiny bit harder to use multiple
times in a single fight.