fix: clean up merge-induced bugs in dark bargain
As well as abort the ability if the servant fails to arrive, instead of
charging the player zombies for nothing.

Unlike other ally abilities, Yred's Dark Bargain is used by players who
have gobs of allies, so many that they might not have any available
space for their servant. Aborting may leak a small amount of
information, but the cases where that happens are either: ones where the
player is getting it any way in virtue of their zombies, or so
vanishingly rare and punishing that its probably ok to leak here.