fix: don't let Usk steal other gods' powers (G-Flex)
As discovered by G-Flex, if you have a religious ability quivered and
convert to Uskayaw, the quivered ability registers as valid and can be
used via quiver commands if you meet its piety threshold. This commit
changes `god_power_usable` to do an explicit god check. To support this
check I refactored the god_powers static to use a construct-on-first-use
paradigm, and added the god field directly to the struct type.