feat: Slightly tone down a float vault (Lici)
For minmay_chambers_of_sevens, the final chamber of sevens has seven 9
out-of-depth monsters, which is a bit steep for earlier mid dungeon
levels like D:7. Allowable, but verging into runed door territory, which
this vault isn't. Replace 3 9s with 0 spawns and downgrade two of the
loot items, preserving the count of seven monsters and seven loot items
in the final chamber.