feat: more control over action panel item order
Previously, item order was determined by enum order. This change uses
the action_panel option itself to determine primary order (this was
already implemented in an ordered way), and item letter for the
suborder. It's possible that the latter will be too confusing by
default, and if so, we can return to subtype as the secondary sort
(which provides a stable ordering across games).

I've also tweaked the default so that it is actually enum order, for
some level of consistency before and after this commit.