feat: action panel glyph mode
I'm not sure this is really that useful, so it's currently not tied to
actually being in glyphs mode and has its own option. We'll see if
people actually use it; if not, (and if I don't find a way to improve
it) it may just be better to remove before release.