Let Yara's Unravelling turn summons into explosions (Undo)
Yara's is a fairly situational spell, so to give it some slightly wider
utility, allow it to also be cast on summoned monsters, unravelling the
magic that summoned them and killing them immediately. This allows it to be
used as a form of single-target abjuration that can additionally damage the
summoner with the resulting explosion.