fix: consistently treat zombie base types
In 1e9075421c5 zombies were de-anonymized. As an unfortunate side
effect, the function mons_zombie_base was changed to return the species
of the base zombie, not the base monster class used to define the
zombie.

As a result, most dramatically noticed for the damage of orcish derived
undead, but also relevant for spriggan defender zombie action eneriges
and any monster with a species much weaker than its type, the zombie was
based on the underlying species.

This commit simplifies some uses of the function and is being made after
verifying every other use will work correctly (or as originally
intended) with mons_zombie_base returning the monster class, and not the
underlying species.

Undead orc warlords should once again be as spooky as they seem from
their name.