refactor: add helper functions for drawing panel actions
The code for drawing actions and item glyphs is a bit complex, so it's
better to split it up into separate functions.

Also, we drew an item glyph and the cursor for some items twice. This
happened if an item had a base tile and a brand icon.

Fix this and simplify handling of `adjusted_scale`.