feat: clicking for inventory item actions
One thing the action panel has revealed is how minimally implemented
clicking is in the webtiles UI. I realize that lots of advanced players
are perfectly fine with never using a mouse, but allowing it makes
things easier for newer players who aren't used to huge keybindings.
This commit enables clicking on the actions in an item description popup
(which is one of my first targets for this kind of change because it's
so obviously missing when you right click from the action panel).

The implementation is entirely client-side: it converts any strings like
"e(v)oke" into a span with the `data-hotkey` attribute set to the hotkey
in parens. This might be described as hacky (perhaps sending structured
action data from the server would be better), but it was really easy to
implement and actually seems pretty robust/useful. Part of the trick was
that someone (aidanh I assume) has previously implemented an amazing
thing where if you just set this data attribute, an on click handler
magically springs into existence that sends the hotkey.