fix: don't rest forever when divine vigour ends (ardl)
If divine vigour ends while resting, bringing the player to full health,
the delay would not end on its own, because the delay end condition was
apparently only checked when increasing hp or mp. (I thought the delay
was also supposed to check, but it seems I've been misreading this
could. Possibly this would be a lot more robust than what we're doing
now...) For most players this doesn't matter, because the divine vigour
end message interrupts the delay. But any player who has something like
`runrest_ignore_message += pray:` can trigger the bug. (Thanks to ardl
for his ...involved rc, I got some good bisecting practice to find
the interaction.)

To fix this, this commit adds a check to calc_hp on top of inc_hp, which
should also cover any similar cases that have so far been missed.

Also, some cleanup of resting checks to be more uniform.