feat: expand out webtiles mouse control
Webtiles mouse control has always been a bit of an incomplete feature; I
think edlothiol only barely started on this in 2012 in 52f5b20b692aa,
which was (afaict) his last commit. No one else has tried to fix this up
since.  This commit fills out the feature set a bit. Previously, you
could use left click to move and attack in command mode, and there was a
minimal tooltip for monsters, and that's it. This commit adds:

1. right click to describe in regular view, targeting view, and map mode.
2. left click to fire in targeting mode. (Or other behaviors, e.g. in
`x` this triggers autotravel.)
3. synchronization of a targeting ray with the mouse cursor position.
This is probably the dodgiest feature in that (a) it may run into issues
with latency (it's handled on the server side) and (b) it may make it
too easy to misclick. The implementation follows local tiles but I
wonder if this would be better with something like, click to select
target, click again to fire. (Also, this was rather involved to
implement as it revealed a subtle bug in the direction chooser.)
4. Slightly less minimal tooltips. This only shows tooltips for monsters
still, because it's a bit constrained by what info there is on the
client side, and constantly doing popups with just the click hints was
pretty annoying in my testing.

One departure from local tiles: mousing over a square out of view
doesn't affect the currently set ray. I found this a lot more natural
and it works better with mousing from an object in the action panel to
the targeting range (which almost always involves crossing out-of-los
squares). Possibly this should be ported to local tiles.

As before, all of this stuff is entirely disabled with
`tile_web_mouse_control = false`.