Revert "fix: don't force picking up of newly IDd equipment"
This reverts commit f5e35ccd160c657ff46174990fc3fe4e97655484.

As implemented it did not have the correct behavior, and was still
incorrect after 99a87601b6. Discussion with players revealed there are
many users who expect autopickup to make its "should pick up decision"
on an item when it comes into view, while others would like their
autopickup options applied with respect to the item
post-floor-identification.

So I'm reverting this to the behavior before f5e35ccd: an item will be
autopickedup if it meets the user's autopickup preferences before it is
floor-identified. The correct solution here is to add a boolean game
option that is checked before setting NEED_AUTOPICKUP_KEY in the
relevant lines (the ones modified by this commit).