feat: revise the ring of the Mage
The ring of the Mage was a not-super exciting Wiz ring with light
bonuses. In keeping with the "good or interesting" unrand revision
paradigm, this commit revises the ring of the Mage to be a Wiz/Archmagi
ring.

Archmagi on an easty to swap slot might raise some eyebrows, but it is
no worse than enhancer swapping. The hope is that this item might
convince some more hybridization, helping make up for both the penalty
of heavy armour and the lack of intelligence of heavy armour users.

Of course, it'll also be an exciting find for any spellcaster who has
body slot Archmagi, thanks to the stacking change. Boom pow!

XXX: The _resist_composer in output.cc assumes that resists with a max
of < 5 should get spacing, if it becomes possible to get 3 pips of
Archmagi this will cause display issues % on console.