feat: chill lofi orbs to ponder/crawl with
A shield-slot analog of scarves, meant to play with the same sort of
defense/interesting ego trade-off. Orbs give no SH and no penalty.

"Caster" and some "battlemage" archetype characters are almost always
steered to shields through a confluence of factors:

- cheap one-handers (demon weapons and shortblades) that can be amplified
  with magic
- one-handed enhancer staves leaving the other hand flapping around
  empty and begging for a shield
- shield penalties are easier for low-strength characters to overcome

Forthcoming commits will introduce orb egos (currently ponderous is a
placeholder but will not be on the final list) aimed at expanding the
options for these characters. Another direction to pursue in this design
space is to create some 2-handers that tempt such characters, this is
not an exclusive direction so both could happen, but for now orbs are
what we've got.