feat: add key hints and hotkeys to feature descriptions
Also, add clicking to webtiles, with some other misc description
enhancements. This turned out to be a mess of special cases, many of
which I've probably still missed. The commands that are supported are
up, down, open, close, and some shortcuts that go directly to the map
view for the other side of explored staircases.

Bugfixes: correctly handle unseen tiles with webtiles right click; fix a
subtle ordering issue in description hotkey  handling.