Store polar vortex power on cast
We historically didn't do that to discourage, say, wielding a staff
of ice, casting the spell, and then unwielding the staff to do something
else. But... usually if you're casting a level 9 ice spell, you sort of
want to have ice enhancers for other things (eg ozo's fridge), and the
un-stored power system encourages other silly swaps - eg start with a
ring of wizardry to get vortex castable, then swap to =ice once it was
running.

Anyway, not storing power meant that using exegesis on vortex was a
nasty player trap. Now that's fixed!