Make Yara's Violent Unravelling smite-targeted
Along similar lines to Inner Flame and its recent change, Unravelling is a
fun hex spell with some good utility, but is often quite fiddly to make
effective use of. Change it to also be smite-targeted, allowing more
precise enchantment/summon removal without the chance of a non-enchanted
creature blocking the beam and making it ineffective.